backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-dotenv
requests
